Feat/Added basket activity
I added a basket view. The cart displays information such as a picture of the dish, name, quantity and price. At the bottom of the view, the total value of the entire cart is displayed, as well as a button to proceed to payment. I also added the ability to remove a dish from the cart by sliding it to the right (and a sliding animation).