Feat/added navigation to checkout
I added navigation to finalize the order. If the shopping cart is empty an appropriate message will be displayed. The order finalization view displays the names of the meals, their quantity, price and total order value. I have also added a form for entering data such as name and a button for payment (the logic for clicking the button will be added later)