Feat/added create order and order details
I added the ability to place an order. After a successful order, its details are displayed, such as id, board number, menu items, date and value of the entire order. After the order is placed, it is saved in shared preferences and the contents of the shopping cart are cleared. The ability to return to the menu view by clicking on the application logo has also been added.