Feat/added order history view
I added an order history that displays the user's last 10 orders. The order history is stored in Shared Preferences under the name “user_orders”. Clicking on any order in the OrderHistory activity takes you to the OrderDetailsActivity. If the user's order history is empty a corresponding message is displayed.
The appearance of BoardActivity has also been improved in this commit.